chore: move docker auth tests to another test file
