+++ docker_test.go
@@ -1073,191 +1073,6 @@ func TestContainerCreationWaitsForLog(t *testing.T) {
 	terminateContainerOnEnd(t, ctx, mysqlC)
 }
 
-func TestBuildContainerFromDockerfile(t *testing.T) {
-	t.Log("getting context")
-	ctx := context.Background()
-	t.Log("got context, creating container request")
-	req := ContainerRequest{
-		FromDockerfile: FromDockerfile{
-			Context: "./testresources",
-		},
-		ExposedPorts: []string{"6379/tcp"},
-		WaitingFor:   wait.ForLog("Ready to accept connections"),
-	}
-
-	redisC, err := prepareRedisImage(ctx, req, t)
-	require.NoError(t, err)
-	terminateContainerOnEnd(t, ctx, redisC)
-}
-
-func TestBuildContainerFromDockerfileWithDockerAuthConfig(t *testing.T) {
-	// using the same credentials as in the Docker Registry
-	base64 := "dGVzdHVzZXI6dGVzdHBhc3N3b3Jk" // testuser:testpassword
-	t.Setenv("DOCKER_AUTH_CONFIG", `{
-		"auths": {
-				"localhost:5000": { "username": "testuser", "password": "testpassword", "auth": "`+base64+`" }
-		},
-		"credsStore": "desktop"
-	}`)
-
-	prepareLocalRegistryWithAuth(t)
-	defer func() {
-		ctx := context.Background()
-		testcontainersClient, err := client.NewClientWithOpts(client.WithVersion(daemonMaxVersion))
-		if err != nil {
-			t.Log("could not create client to cleanup registry: ", err)
-		}
-
-		_, err = testcontainersClient.ImageRemove(ctx, "localhost:5000/redis:5.0-alpine", types.ImageRemoveOptions{
-			Force:         true,
-			PruneChildren: true,
-		})
-		if err != nil {
-			t.Log("could not remove image: ", err)
-		}
-
-	}()
-
-	t.Log("getting context")
-	ctx := context.Background()
-	t.Log("got context, creating container request")
-	req := ContainerRequest{
-		FromDockerfile: FromDockerfile{
-			Context:    "./testresources",
-			Dockerfile: "auth.Dockerfile",
-		},
-
-		ExposedPorts: []string{"6379/tcp"},
-		WaitingFor:   wait.ForLog("Ready to accept connections"),
-	}
-
-	redisC, err := prepareRedisImage(ctx, req, t)
-	require.NoError(t, err)
-	terminateContainerOnEnd(t, ctx, redisC)
-}
-
-func TestBuildContainerFromDockerfileShouldFailWithWrongDockerAuthConfig(t *testing.T) {
-	// using different credentials than in the Docker Registry
-	base64 := "Zm9vOmJhcg==" // foo:bar
-	t.Setenv("DOCKER_AUTH_CONFIG", `{
-		"auths": {
-			"localhost:5000": { "username": "foo", "password": "bar", "auth": "`+base64+`" }
-		},
-		"credsStore": "desktop"
-	}`)
-
-	prepareLocalRegistryWithAuth(t)
-
-	t.Log("getting context")
-	ctx := context.Background()
-	t.Log("got context, creating container request")
-	req := ContainerRequest{
-		FromDockerfile: FromDockerfile{
-			Context:    "./testresources",
-			Dockerfile: "auth.Dockerfile",
-		},
-		ExposedPorts: []string{"6379/tcp"},
-		WaitingFor:   wait.ForLog("Ready to accept connections"),
-	}
-
-	redisC, err := prepareRedisImage(ctx, req, t)
-	require.Error(t, err)
-	terminateContainerOnEnd(t, ctx, redisC)
-}
-
-func TestCreateContainerFromPrivateRegistry(t *testing.T) {
-	// using the same credentials as in the Docker Registry
-	base64 := "dGVzdHVzZXI6dGVzdHBhc3N3b3Jk" // testuser:testpassword
-	t.Setenv("DOCKER_AUTH_CONFIG", `{
-		"auths": {
-				"localhost:5000": { "username": "testuser", "password": "testpassword", "auth": "`+base64+`" }
-		},
-		"credsStore": "desktop"
-	}`)
-
-	prepareLocalRegistryWithAuth(t)
-
-	ctx := context.Background()
-	req := ContainerRequest{
-		Image:           "localhost:5000/redis:5.0-alpine",
-		AlwaysPullImage: true, // make sure the authentication takes place
-		ExposedPorts:    []string{"6379/tcp"},
-		WaitingFor:      wait.ForLog("Ready to accept connections"),
-	}
-
-	redisContainer, err := GenericContainer(ctx, GenericContainerRequest{
-		ContainerRequest: req,
-		Started:          true,
-	})
-	require.Nil(t, err)
-	terminateContainerOnEnd(t, ctx, redisContainer)
-}
-
-func prepareLocalRegistryWithAuth(t *testing.T) {
-	ctx := context.Background()
-	wd, err := os.Getwd()
-	assert.NoError(t, err)
-	req := ContainerRequest{
-		Image:        "registry:2",
-		ExposedPorts: []string{"5000:5000/tcp"},
-		Env: map[string]string{
-			"REGISTRY_AUTH":                             "htpasswd",
-			"REGISTRY_AUTH_HTPASSWD_REALM":              "Registry",
-			"REGISTRY_AUTH_HTPASSWD_PATH":               "/auth/htpasswd",
-			"REGISTRY_STORAGE_FILESYSTEM_ROOTDIRECTORY": "/data",
-		},
-		Mounts: ContainerMounts{
-			ContainerMount{
-				Source: GenericBindMountSource{
-					HostPath: fmt.Sprintf("%s/testresources/auth", wd),
-				},
-				Target: "/auth",
-			},
-			ContainerMount{
-				Source: GenericBindMountSource{
-					HostPath: fmt.Sprintf("%s/testresources/data", wd),
-				},
-				Target: "/data",
-			},
-		},
-		WaitingFor: wait.ForExposedPort(),
-	}
-
-	genContainerReq := GenericContainerRequest{
-		ProviderType:     providerType,
-		ContainerRequest: req,
-		Started:          true,
-	}
-
-	t.Log("creating registry container")
-
-	registryC, err := GenericContainer(ctx, genContainerReq)
-	assert.NoError(t, err)
-
-	t.Cleanup(func() {
-		assert.NoError(t, registryC.Terminate(context.Background()))
-	})
-
-	_, cancel := context.WithCancel(context.Background())
-	t.Cleanup(cancel)
-}
-
-func prepareRedisImage(ctx context.Context, req ContainerRequest, t *testing.T) (Container, error) {
-	genContainerReq := GenericContainerRequest{
-		ProviderType:     providerType,
-		ContainerRequest: req,
-		Started:          true,
-	}
-
-	t.Log("creating redis container")
-
-	redisC, err := GenericContainer(ctx, genContainerReq)
-
-	t.Log("created redis container")
-
-	return redisC, err
-}
-
 func Test_BuildContainerFromDockerfileWithBuildArgs(t *testing.T) {
 	t.Log("getting ctx")
 	ctx := context.Background()
